separate testing file to avoid mpi issue
